Implement can.getName symbol behavior
 Ref: https://github.com/canjs/can-symbol/issues/16
