+++ package.json
@@ -41,7 +41,7 @@
     "can-observation": "^3.3.4",
     "can-reflect": "^1.2.1",
     "can-simple-observable": "^1.0.0",
-    "can-symbol": "^1.0.0",
+    "can-symbol": "^1.3.0",
     "can-types": "^1.1.0",
     "can-util": "^3.9.0"
   },
